epub-maker: extract page grid drag and drop hook
